pyproject.toml
# https://packaging.python.org/en/latest/guides/writing-pyproject-toml/
[project]
name = "transcribe"
requires-python = ">=3.11"
readme = "README.md"
version = "0.1.0"
# https://github.com/yt-dlp/yt-dlp/releases
# curl-cffi is for the impersonnation: https://github.com/yt-dlp/yt-dlp#impersonation
dependencies = [
  "yt-dlp[default,curl-cffi]==2026.03.17",
  "webvtt-py==0.5.1",
  "packaging==26.0",
  "mcp[cli]==1.27.0",
  "httpx==0.28.1",
  "typer==0.24.1",
  "google-auth==2.49.2",
]

[dependency-groups]
dev = [
  "inline-snapshot>=0.32.6",
  "pytest==9.0.3",
  "pytest-asyncio==1.3.0",
]

# Built backend
# https://packaging.python.org/en/latest/guides/writing-pyproject-toml/#declaring-the-build-backend
[build-system]
# no setuptools, no *.egg-info/
# requires = ["setuptools"]
# build-backend = "setuptools.build_meta"
requires = ["hatchling"]
build-backend = "hatchling.build"

# Executable script
# https://packaging.python.org/en/latest/guides/writing-pyproject-toml/#creating-executable-scripts
[project.scripts]
# Note the hyphen vs underscore —
# the CLI command uses a hyphen (as you named it),
# but the Python module name uses an underscore since hyphens aren't valid in Python identifiers.
transcribe = "gerardnico.transcribe.cli:cli"

# To resolve: raceback (most recent call last):
#  File "/home/admin/.local/bin/transcript-downloader", line 3, in <module>
#    from transcript_downloader import main
#   ModuleNotFoundError: No module named 'transcript_downloader'
#   The behavior difference between laptops is likely a setuptools version difference —
#   older setuptools used to auto-discover top-level .py modules, newer versions require explicit py-modules declaration.
# [tool.setuptools]
# py-modules = ["transcript_downloader"]

# Package find with hatch
[tool.hatch.build.targets.wheel]
packages = ["src/gerardnico"]

# Package find with setuptools:
[tool.setuptools.packages.find]
where = ["src"]

# Test
[tool.pytest.ini_options]
asyncio_mode = "auto"        # enables automatic async detection
testpaths = ["tests"]        # where pytest looks
pythonpath = ["src"]         # add src/ to sys.path (src layout)
python_files = ["test_*.py"]
